Added message when no records are found on a overview page
