+++ DemaWare.DemaIdentify.Web/Pages/Admin/Identity/Role/RoleDelete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/admin/identity/role/{roleId:Guid}/delete"
-@model DemaWare.DemaIdentify.Web.Pages.Admin.Identity.RoleDeleteModel
+@model DemaWare.DemaIdentify.Web.Pages.Admin.Identity.Role.RoleDeleteModel
 @inject LocalizationService localizer
 
 @{
