+++ DemaWare.DemaIdentify.Web/Pages/Admin/Identity/Role/RoleOverview.cshtml.cs
@@ -3,7 +3,7 @@ using DemaWare.General.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace DemaWare.DemaIdentify.Web.Pages.Admin.Identity; 
+namespace DemaWare.DemaIdentify.Web.Pages.Admin.Identity.Role; 
 public class RoleOverviewModel : PageModel {
     private readonly IdentityService _identityService;
