+++ DemaWare.DemaIdentify.Web/Pages/Admin/Identity/User/UserOverview.cshtml.cs
@@ -2,7 +2,7 @@ using DemaWare.DemaIdentify.BusinessLogic.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace DemaWare.DemaIdentify.Web.Pages.Admin.Identity; 
+namespace DemaWare.DemaIdentify.Web.Pages.Admin.Identity.User; 
 public class UserOverviewModel : PageModel {
 	private readonly IdentityService _identityService;
