+++ DemaWare.DemaIdentify.Web/Pages/Admin/Identity/User/UserRoleDelete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/admin/identity/user/{userId:Guid}/role/{roleName}/delete"
-@model DemaWare.DemaIdentify.Web.Pages.Admin.Identity.UserRoleDeleteModel
+@model DemaWare.DemaIdentify.Web.Pages.Admin.Identity.User.UserRoleDeleteModel
 @inject LocalizationService localizer
 
 @{
