+++ DemaWare.DemaIdentify.Web/Pages/Admin/Identity/User/UserRoles.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/admin/identity/user/{userId:Guid}/role"
-@model DemaWare.DemaIdentify.Web.Pages.Admin.Identity.UserRolesModel
+@model DemaWare.DemaIdentify.Web.Pages.Admin.Identity.User.UserRolesModel
 @inject LocalizationService localizer
 
 @{
