+++ DemaWare.DemaIdentify.Web/Pages/Admin/Template/Email/EmailCreate.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/admin/template/email/create"
-@model DemaWare.DemaIdentify.Web.Pages.Admin.Template.EmailCreateModel
+@model DemaWare.DemaIdentify.Web.Pages.Admin.Template.Email.EmailCreateModel
 @inject LocalizationService localizer
 
 @{
