+++ DemaWare.DemaIdentify.Web/Pages/Admin/Template/Email/EmailDelete.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/admin/template/email/{templateId:Guid}/delete"
-@model DemaWare.DemaIdentify.Web.Pages.Admin.Template.EmailDeleteModel
+@model DemaWare.DemaIdentify.Web.Pages.Admin.Template.Email.EmailDeleteModel
 @inject LocalizationService localizer
 
 @{
