+++ DemaWare.DemaIdentify.Web/Pages/Admin/Template/Email/EmailDelete.cshtml.cs
@@ -3,7 +3,7 @@ using DemaWare.DemaIdentify.Models.Template;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace DemaWare.DemaIdentify.Web.Pages.Admin.Template; 
+namespace DemaWare.DemaIdentify.Web.Pages.Admin.Template.Email; 
 public class EmailDeleteModel : PageModel {
 	private readonly TemplateService _templateService;
