+++ DemaWare.DemaIdentify.Web/Pages/Admin/Template/Email/EmailEdit.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/admin/template/email/{templateId:Guid}/edit"
-@model DemaWare.DemaIdentify.Web.Pages.Admin.Template.EmailEditModel
+@model DemaWare.DemaIdentify.Web.Pages.Admin.Template.Email.EmailEditModel
 @inject LocalizationService localizer
 
 @{
