+++ DemaWare.DemaIdentify.Web/Pages/Shared/_AdminLayout.cshtml
@@ -32,12 +32,12 @@
 								<li class="nav-item dropdown">
 									<a href="#" class="nav-link dropdown-toggle" id="navbarDropdownMenuIdentityLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">@localizer.GetLocalizedHtmlString("Identities")</a>
 									<ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuIdentityLink">
-										<li><a asp-page="/Admin/Identity/UserOverview" class="dropdown-item">@localizer.GetLocalizedHtmlString("Users")</a></li>
-										<li><a asp-page="/Admin/Identity/RoleOverview" class="dropdown-item">@localizer.GetLocalizedHtmlString("Roles")</a></li>
+										<li><a asp-page="/Admin/Identity/User/UserOverview" class="dropdown-item">@localizer.GetLocalizedHtmlString("Users")</a></li>
+										<li><a asp-page="/Admin/Identity/Role/RoleOverview" class="dropdown-item">@localizer.GetLocalizedHtmlString("Roles")</a></li>
 									</ul>
 								</li>
 								<li class="nav-item"><a asp-page="/Admin/Organisation/OrganisationOverview" class="nav-link">@localizer.GetLocalizedHtmlString("Organisations")</a></li>
-								<li class="nav-item"><a asp-page="/Admin/Template/EmailOverview" class="nav-link">@localizer.GetLocalizedHtmlString("Templates")</a></li>
+								<li class="nav-item"><a asp-page="/Admin/Template/Email/EmailOverview" class="nav-link">@localizer.GetLocalizedHtmlString("Templates")</a></li>
 								<li class="nav-item"><a asp-page="/Logout" class="nav-link">@localizer.GetLocalizedHtmlString("Logout")</a></li>
 							</ul>
 						</div>
